Add files via upload
two certificates created. one valid and one which has got an invalid common name.
in the need of broken certificates for the common name 'localhost'.
can begin creating test for different libraries, but only some tests are possible at the moment, the same codes will most likely still work in the future, maybe some little modifications will be needed but when the codes do already exist they are much easier to modify.
ports 100-X are reserved for the different certificates at the moment.
more info will be coming and does in fact already exist in the readMe:s in the files uploaded.
you can build and run the current project in the tryTLS/ssltest/https folder after unzipping the file uploaded, read the readMe:s for help if you need any.