Updated the Copyright note in the LICENSE
Based on the quick feed trying to clarify Copyright so that
it doesn't imply assigning the copyrights if you contribute.

Currently policy is that you indicate or imply to distribute your
contribution under tha project license or attach the license of
your choice to your contribution. if your license is compatible
with the main license the this project can be distributed and
used whole.

Further Copyright transfer & assignment is unnecessary and this
I propose that misleading Copyright notice to be corrected. I
recommend alwayd distributing the project in a form that either
includes backreference to the project repository or at least
identifies the list of individual contributors so that
original Copyrights can be backtracked if necessary and
parental right gets respected.