bug: fix Python 3.8+ linux compat.
Prior to this commit the `runners/trytls/utils.py` script would fail to
execute on Python 3.8+ systems. This is because the deprecated
`platform.linux_distribution` func used by this script was removed in
Python 3.8.

This commit updates the script to try to conditionally use the `distro`
module[1] as a replacement. The `setup.py` script is updated to
conditionally require this dependency on systems using Python 3.8+.

This change is sufficient to get trytls working on my Python 3.10
system.

[0]: https://docs.python.org/3.7/library/platform.html#platform.linux_distribution
[1]: https://pypi.org/project/distro/